requirements.txt
asgiref==3.8.1
Django==5.2
django-modeltranslation==0.19.14
pillow==11.2.1
sqlparse==0.5.3
